backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
pydantic==2.5.0
pyyaml==6.0.1
aiofiles==23.2.1
httpx==0.25.1
ultralytics==8.0.220
torch==2.1.0
python-dotenv==1.0.0
